Copy over split association work
This copies over the functionality and gets one type of test passing.

Next:

1) Get all tests copied over from the gem so we know everything is
covered
2) Clean up existing code to make it less gross (this gem is pretty
gross). A lot of it is a giant monkey patch and I think some of the code
is more duplicated than necessary.
3) Decide on the API - I'm not sure `split: true` is going to
fly with DHH.

Later:

I'm not sure has_one associations work with this gem. There is also the
whole transactions across clusters issue but I think that's way out of
scope for this gem.